upgrade to hull-node 0.11.0-beta.3
