+++ server/index.js
@@ -1,7 +1,5 @@
+import librato from "librato-node";
 import Hull from "hull";
-import { Agent } from "./agent";
-import { Server } from "./server";
-import config from "./config";
 import { name } from "../manifest.json";
 
 if (process.env.LOGSTASH_HOST && process.env.LOGSTASH_PORT) {
@@ -16,4 +14,16 @@ if (process.env.LOGSTASH_HOST && process.env.LOGSTASH_PORT) {
   Hull.logger.info("start", { transport: "console" });
 }
 
-export default { Agent, Server, config };
+if (process.env.LIBRATO_TOKEN && process.env.LIBRATO_USER) {
+  librato.configure({
+    email: process.env.LIBRATO_USER,
+    token: process.env.LIBRATO_TOKEN
+  });
+  librato.on("error", (err) => {
+    console.error(err);
+  });
+  process.once("SIGINT", () => {
+    librato.stop(); // stop optionally takes a callback
+  });
+  librato.start();
+}
