+++ server/server.js
@@ -1,11 +1,8 @@
 import _ from "lodash";
-import express from "express";
 import cors from "cors";
-import path from "path";
-import { Hull, NotifHandler, BatchHandler, Middleware, OAuthHandler } from "hull";
-import { Strategy } from "passport-forcedotcom";
 import librato from "librato-node";
-import { renderFile } from "ejs";
+import { notifHandler, batchHandler, oAuthHandler } from "hull/lib/utils";
+import { Strategy } from "passport-forcedotcom";
 
 import BatchSyncHandler from "./batch-sync";
 import Agent from "./agent";
@@ -21,32 +18,13 @@ function save(hull, ship, settings) {
 }
 
 
-export default function Server({ hostSecret }) {
-  if (process.env.LIBRATO_TOKEN && process.env.LIBRATO_USER) {
-    librato.configure({
-      email: process.env.LIBRATO_USER,
-      token: process.env.LIBRATO_TOKEN
-    });
-    // librato.on("error", () => {
-    //   console.error(err);
-    // });
+module.exports = function Server(app, options = {}) {
+  const { Hull, hostSecret, port, onMetric, clientConfig = {} } = options;
+  const connector = new Hull.Connector({ hostSecret, port, clientConfig });
 
-    process.once("SIGINT", () => {
-      librato.stop(); // stop optionally takes a callback
-    });
+  connector.setupApp(app);
 
-    librato.start();
-  }
-
-  const app = express();
-
-  app.set("views", `${__dirname}/../views`);
-  app.set("view engine", "ejs");
-  app.engine("html", renderFile);
-  app.use(express.static(path.resolve(__dirname, "..", "dist")));
-  app.use(express.static(path.resolve(__dirname, "..", "assets")));
-
-  app.use("/auth", OAuthHandler({
+  app.use("/auth", oAuthHandler({
     hostSecret,
     name: "Salesforce",
     Strategy,
@@ -55,7 +33,7 @@ export default function Server({ hostSecret }) {
       clientSecret: process.env.CLIENT_SECRET, // Client Secret
       scope: ["refresh_token", "api"] // App Scope
     },
-    isSetup(req, { /* hull,*/ ship }) {
+    isSetup(req, { hull, ship }) {
       if (!!req.query.reset) return Promise.reject();
       const { access_token, refresh_token, instance_url } = ship.private_settings || {};
 
@@ -84,7 +62,7 @@ export default function Server({ hostSecret }) {
     },
   }));
 
-  app.post("/sync", Middleware({ hostSecret }), (req, res) => {
+  app.post("/sync", connector.clientMiddleware(), (req, res) => {
     const { client: hull, ship } = req.hull;
     Agent.fetchChanges(hull, ship).then((result) => {
       res.json({ ok: true, result });
@@ -95,7 +73,7 @@ export default function Server({ hostSecret }) {
     });
   });
 
-  app.post("/fetch-all", Middleware({ hostSecret }), (req, res) => {
+  app.post("/fetch-all", connector.clientMiddleware(), (req, res) => {
     const { client: hull, ship } = req.hull;
     Agent.fetchAll(hull, ship).then((result) => {
       res.json({ ok: true, result });
@@ -106,9 +84,13 @@ export default function Server({ hostSecret }) {
     });
   });
 
-  app.post("/notify", NotifHandler({
+  app.post("/notify", notifHandler({
+    userHandlerOptions: {
+      groupTraits: false,
+      maxSize: 1,
+      maxTime: 1
+    },
     hostSecret,
-    groupTraits: false,
     onSusbscribe(message, context) {
       Hull.logger.warn("Hello new subscriber !", { message, context });
     },
@@ -131,7 +113,7 @@ export default function Server({ hostSecret }) {
     }
   }));
 
-  app.post("/batch", BatchHandler({
+  app.post("/batch", batchHandler({
     hostSecret,
     batchSize: 2000,
     groupTraits: false,
@@ -144,11 +126,7 @@ export default function Server({ hostSecret }) {
     }
   }));
 
-  app.get("/manifest.json", (req, res) => {
-    res.sendFile(path.resolve(__dirname, "..", "manifest.json"));
-  });
-
-  app.get("/schema(/:type)", cors(), Middleware({ hostSecret, requireCredentials: false }), (req, res) => {
+  app.get("/schema(/:type)", cors(), connector.clientMiddleware({ requireCredentials: false }), (req, res) => {
     const { type } = req.params || {};
     const { client: hull, ship } = req.hull;
     return Agent.getFieldsSchema(hull, ship).then((definitions = {}) => {
@@ -161,8 +139,7 @@ export default function Server({ hostSecret }) {
     });
   });
 
-  return {
-    listen: port => app.listen(port),
-    exit: () => BatchSyncHandler.exit()
-  };
-}
+  connector.startApp(app);
+
+  return app;
+};
